Display confirmation message if a user attempts to delete a non-empty job. Introduced factory messages for working with confirmation option panes.

+++ ute/src/main/java/net/lmxm/ute/gui/MainFrame.java
@@ -92,7 +92,9 @@ import net.lmxm.ute.beans.tasks.Task;
 import net.lmxm.ute.configuration.ConfigurationReader;
 import net.lmxm.ute.configuration.ConfigurationUtils;
 import net.lmxm.ute.configuration.ConfigurationWriter;
+import net.lmxm.ute.gui.components.GuiComponentConfirmation;
 import net.lmxm.ute.gui.components.MainTree;
+import net.lmxm.ute.gui.components.OptionPaneFactory;
 import net.lmxm.ute.gui.components.StatusOutputPanel;
 import net.lmxm.ute.gui.components.StatusOutputTab;
 import net.lmxm.ute.gui.dialogs.AboutDialog;
@@ -408,6 +410,16 @@ public final class MainFrame extends JFrame implements ConfigurationHolder, Acti
 		}
 
 		final Job job = (Job) userObject;
+
+		if (job.getTasks().size() > 0) {
+			final int result = OptionPaneFactory.showConfirmation(this, GuiComponentConfirmation.DELETE_JOB,
+					job.getId(), job.getTasks().size());
+
+			if (result == JOptionPane.NO_OPTION) {
+				return;
+			}
+		}
+
 		configuration.getJobs().remove(job);
 		mainTree.deleteJob(job);
 	}
